% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_eta.R
\name{analytical_eta}
\alias{analytical_eta}
\title{General function to calculate eta analitically for gaussian and gamma distributions}
\usage{
analytical_eta(
  mesh,
  case = c("gaussian", "lognormal", "gamma"),
  param_1_y,
  param_2_y,
  param_1_x,
  param_2_x
)
}
\arguments{
\item{param_1_y}{Either mu or shape (gaussian / gamma) of the non-diseased population.}

\item{param_2_y}{Either stdev or rate (gaussian / gamma) of the non-diseased population.}

\item{param_1_x}{Either mu or shape (gaussian / gamma) of the diseased population.}

\item{param_2_x}{Either stdev or rate (gaussian / gamma) of the diseased population.}

\item{mux}{A numeric value for the mean of the distribution (e.g., mean for Normal, Log-Normal).}

\item{mesh_size}{The number of intervals to use for the mesh/grid (default is m).}
}
\value{
The calculated eta value.
}
\description{
This function allows calculating eta using custom distributions and their respective transformations.
}
