% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_box_cox}
\alias{apply_box_cox}
\title{Perform Box-Cox transformation on two variables}
\usage{
apply_box_cox(x, y, print_lambda = FALSE)
}
\arguments{
\item{x}{A numeric vector (first variable).}

\item{y}{A numeric vector (second variable).}

\item{print_lambda}{A logical value; if TRUE, the lambda value (Box-Cox transformation parameter) is printed.}
}
\value{
A list containing the transformed vectors \link{transformed_x, transformed_y}.
}
\description{
This function applies the Box-Cox transformation to two numeric vectors. If any values are less than or equal to zero,
a small constant is added to both vectors to make them positive.
}
