% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bias}
\alias{get_bias}
\title{Calculate the bias of the estimator}
\usage{
get_bias(estimates, true_value)
}
\arguments{
\item{estimates}{A numeric vector of estimated values.}

\item{true_value}{The true value of the estimator.}
}
\value{
The bias of the estimator (difference between mean of estimates and true value).
}
\description{
Calculate the bias of the estimator
}
