% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mux}
\alias{get_mux}
\title{Calculate mux value based on (6) from Faraggi, D., & Reiser, B. (2002). Estimation of the area under the ROC curve. Statistics in medicine, 21(20), 3093-3106.}
\usage{
get_mux(auc, sigma_x, sigma_y, mu_y)
}
\arguments{
\item{auc}{The area under the curve (AUC) value.}

\item{sigma_x}{Standard deviation of X.}

\item{sigma_y}{Standard deviation of Y.}

\item{mu_y}{The mean of Y.}
}
\value{
The calculated mux value.
}
\description{
Calculate mux value based on (6) from Faraggi, D., & Reiser, B. (2002). Estimation of the area under the ROC curve. Statistics in medicine, 21(20), 3093-3106.
}
