% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mux_bisection}
\alias{get_mux_bisection}
\title{Estimate the mean parameter from a gaussian that minimizes the AUC difference}
\usage{
get_mux_bisection(
  y_population,
  stdev_x,
  auc_target,
  tol = 0.01,
  max_iter = 100,
  lower = 0,
  upper = 10
)
}
\arguments{
\item{y_population}{A numeric vector of observed values for the population Y.}

\item{stdev_x}{The sigma parameter for the population X.}

\item{auc_target}{The target AUC value to match.}

\item{tol}{The tolerance level for AUC difference (default is 0.01).}

\item{max_iter}{The maximum number of iterations for the bisection method (default is 100).}

\item{lower}{The lower bound used to start the algorithm.}

\item{upper}{The upper bound used to start the algorithm.}
}
\value{
The estimated mean or FALSE if the target AUC is not reached within the maximum iterations.
}
\description{
This function uses the bisection method to find the gaussian distribution mean that minimizes the difference between
the computed AUC and the target AUC.
}
