% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rmse}
\alias{get_rmse}
\title{Calculate the Root Mean Square Error (RMSE)}
\usage{
get_rmse(estimates, true_value)
}
\arguments{
\item{estimates}{A numeric vector of estimated values.}

\item{true_value}{The true value of the estimator.}
}
\value{
The RMSE, a measure of the difference between estimated and true values.
}
\description{
Calculate the Root Mean Square Error (RMSE)
}
