% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_eta.R
\name{kernel_density_estimation}
\alias{kernel_density_estimation}
\title{Kernel Density Estimation}
\usage{
kernel_density_estimation(data, points, bandwidth)
}
\arguments{
\item{data}{A numeric vector of data points.}

\item{points}{A numeric vector of points at which to estimate the density.}

\item{bandwidth}{The bandwidth (smoothing parameter) for the Kernel.}
}
\value{
A numeric vector of the estimated density values at the specified points.
}
\description{
This function estimates the probability density function using Kernel density estimation.
}
