% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_eta.R
\name{kernel_eta}
\alias{kernel_eta}
\title{Kernel Eta Estimation}
\usage{
kernel_eta(
  controls,
  cases,
  method,
  t0 = 1,
  mesh_size_kernel = 1000,
  box_cox = FALSE
)
}
\arguments{
\item{controls}{A numeric vector of data points representing the controls sample.}

\item{cases}{A numeric vector of data points representing the cases sample.}

\item{method}{A character string specifying the bandwidth selection method. Options are 'optimal' or 'hscv'. The latter coming from the ks package.}

\item{mesh_size_kernel}{The number of mesh points used for the estimation. Default is 1000.}

\item{box_cox}{A logical value indicating whether to apply a Box-Cox transformation. Default is FALSE.}
}
\value{
A numeric vector of the estimated eta values at the specified points in the mesh.
}
\description{
This function estimates the population eta using Kernel density estimation for controls and cases samples.
It can apply a Box-Cox transformation if specified.
}
\examples{
# Example usage without Box-Cox transformation
controls = rnorm(1000, mean = 0, sd = 1)
cases = rnorm(1000, mean = 1, sd = 1)
eta = kernel_eta(controls, cases)

# Example usage with Box-Cox transformation
eta_boxcox = kernel_eta(controls, cases, box_cox = TRUE)

}
