% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_eta.R
\name{eta_from_roc_curves}
\alias{eta_from_roc_curves}
\title{Calculate eta value from ROC curves}
\usage{
eta_from_roc_curves(roc, roc_dx, mesh)
}
\arguments{
\item{roc}{A numeric vector of ROC curve values.}

\item{roc_dx}{A numeric vector of ROC curve derivatives (roc_prima).}

\item{mesh}{A numeric vector representing the mesh/grid used for integration.}
}
\value{
The standardized eta value.
}
\description{
This function calculates the eta value using the given ROC curves and mesh. The eta is computed based on the formula
that involves the differences between the ROC curve values and mesh intervals.
}
