% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_eta.R
\name{evaluate_kernel_estimation}
\alias{evaluate_kernel_estimation}
\title{Evaluate Kernel Estimation at a Point}
\usage{
evaluate_kernel_estimation(point, function_values, mesh)
}
\arguments{
\item{point}{A numeric value for the point at which to evaluate the function.}

\item{function_values}{A numeric vector of the estimated function values (from Kernel estimation).}

\item{mesh}{A numeric vector of the mesh/grid used in the estimation.}
}
\value{
The evaluated value at the specified point.
}
\description{
This function evaluates the Kernel estimation at a specific point based on the mesh.
}
