% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_eta.R
\name{inverse_kernel_estimation}
\alias{inverse_kernel_estimation}
\title{Find Quantile of Kernel Estimation}
\usage{
inverse_kernel_estimation(point, function_values, mesh)
}
\arguments{
\item{point}{A numeric value representing the quantile to find.}

\item{function_values}{A numeric vector of the estimated function values (from Kernel estimation).}

\item{mesh}{A numeric vector of the mesh/grid used in the estimation.}
}
\value{
The quantile value corresponding to the specified point in the mesh.
}
\description{
This function finds the quantile of the Kernel estimation at a specific point.
}
