% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_eta.R
\name{kernel_distribution_estimation}
\alias{kernel_distribution_estimation}
\title{Kernel Distribution Estimation}
\usage{
kernel_distribution_estimation(data, points, bandwidth)
}
\arguments{
\item{data}{A numeric vector of data points.}

\item{points}{A numeric vector of points at which to estimate the distribution.}

\item{bandwidth}{The bandwidth (smoothing parameter) for the Kernel.}
}
\value{
A numeric vector of the estimated distribution values at the specified points.
}
\description{
This function estimates the cumulative distribution function (CDF) using Kernel density estimation.
}
